requirements.txt
torch>=1.10.0
numpy>=1.20.0
pytest>=7.0.0
scikit-learn>=1.0.0
pyyaml>=5.4.0
tensorboard>=2.8.0
matplotlib>=3.5.0
seaborn>=0.11.0
wandb>=0.12.0
jupyter>=1.0.0
ipython>=7.30.0